chore: fix fixtures for new deletedStaleLineCount field, split acceptance-gate tests out (issue 412)
- orders-http.integration.test.ts / scheduler-jobs.integration.test.ts:
  add the new required OrdersIngestResult.deletedStaleLineCount field
  to their mock "accepted" fixtures (typecheck was red without this).
- orders-ingest.integration.test.ts hit eslint max-lines: 400 after the
  new reconciliation test — moved the acceptance-gate tests (empty
  export, missing column, malformed row, trust-on-first-use, drastic
  drop) into orders-ingest-acceptance.integration.test.ts, same split
  pattern as orders-ingest-posta-fields.integration.test.ts.
